add a new hook to synchronize rows
