+++ packages/portal-proto/src/components/Summary/SummaryCard.tsx
@@ -1,3 +1,4 @@
+import React from "react";
 import { HeaderTitle } from "@/components/tailwindComponents";
 import { HorizontalTable, HorizontalTableProps } from "../HorizontalTable";
 
@@ -5,25 +6,30 @@ export interface SummaryCardProps {
   readonly title?: string;
   readonly tableData: HorizontalTableProps["tableData"];
   readonly customDataTestID?: string;
+  readonly enableSync?: boolean;
+  readonly ref?: React.Ref<HTMLTableElement>;
 }
 
-export const SummaryCard = ({
-  title = "Summary",
-  tableData,
-  customDataTestID,
-}: SummaryCardProps): JSX.Element => {
-  return (
-    <div className="flex flex-col flex-grow">
-      {title !== "" ? (
-        <HeaderTitle>{title}</HeaderTitle>
-      ) : (
-        <div className="h-7 mb-2" />
-      )}
+export const SummaryCard = React.forwardRef<HTMLTableElement, SummaryCardProps>(
+  (
+    { title = "Summary", tableData, customDataTestID, enableSync = false },
+    ref,
+  ): JSX.Element => {
+    return (
+      <div className="flex flex-col gap-2 flex-grow">
+        {title !== "" ? (
+          <HeaderTitle>{title}</HeaderTitle>
+        ) : (
+          <div className="h-7" />
+        )}
 
-      <HorizontalTable
-        customDataTestID={customDataTestID}
-        tableData={tableData}
-      />
-    </div>
-  );
-};
+        <HorizontalTable
+          customDataTestID={customDataTestID}
+          tableData={tableData}
+          enableSync={enableSync}
+          ref={ref}
+        />
+      </div>
+    );
+  },
+);
