+++ packages/portal-proto/src/features/projects/ProjectView.tsx
@@ -1,4 +1,4 @@
-import React from "react";
+import React, { useRef } from "react";
 import { AnnotationDefaults, GdcApiData, ProjectDefaults } from "@gff/core";
 import { FaUser, FaFile, FaEdit } from "react-icons/fa";
 import { SummaryHeader } from "@/components/Summary/SummaryHeader";
@@ -18,6 +18,7 @@ import useScrollToHash from "@/hooks/useScrollToHash";
 import { useViewportSize } from "@mantine/hooks";
 import { LG_BREAKPOINT } from "src/utils";
 import SummaryHeaderControls from "./SummaryHeaderControls";
+import { useSynchronizedRowHeights } from "@/components/HorizontalTable/useSynchronizedRowHeights";
 
 export interface ProjectViewProps extends ProjectDefaults {
   readonly annotation: GdcApiData<AnnotationDefaults>;
@@ -30,6 +31,9 @@ export const ProjectView: React.FC<ProjectViewProps> = (
 ) => {
   const { width } = useViewportSize();
   useScrollToHash(["annotations"]);
+  const leftSummaryTableRef = useRef<HTMLTableElement>(null);
+  const rightSummaryTableRef = useRef<HTMLTableElement>(null);
+  useSynchronizedRowHeights([leftSummaryTableRef, rightSummaryTableRef]);
 
   const Cases = (
     <span className="flex items-center gap-1">
@@ -126,11 +130,17 @@ export const ProjectView: React.FC<ProjectViewProps> = (
           <div className="basis-full lg:basis-1/2">
             <HorizontalTable
               tableData={width >= LG_BREAKPOINT ? leftColumnData : summaryData}
+              ref={leftSummaryTableRef}
+              enableSync={true}
             />
           </div>
           {width >= LG_BREAKPOINT && (
             <div className="basis-1/2 h-full">
-              <HorizontalTable tableData={rightColumnData} />
+              <HorizontalTable
+                tableData={rightColumnData}
+                ref={rightSummaryTableRef}
+                enableSync={true}
+              />
             </div>
           )}
         </div>
